requirements.txt
anthropic>=0.21.0
requests>=2.31.0
python-dotenv>=1.0.0
googlesearch-python>=1.2.3
streamlit>=1.28.0